Modernize the classes setup
