+++ package.json
@@ -29,7 +29,6 @@
     "grunt-contrib-nodeunit": "~0.4.0",
     "grunt-contrib-uglify": "~0.9.0",
     "grunt-contrib-watch": "~0.6.0",
-    "grunt-phpunit": "~0.3.0",
     "grunt-postcss": "~0.6.0",
     "grunt-text-replace": "~0.4.0",
     "grunt-wp-deploy": "~1.1.0",
